refactored problem/parts structure & moved execution from main()
